+++ src/components/mainPage/VideoCard.module.scss
@@ -18,4 +18,5 @@
 .listicon {
   width: 30px;
   height: 30px;
+  color: var(--color-white);
 }
