+++ src/components/mainPage/VideoCards.module.scss
@@ -1,0 +1,6 @@
+.top {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  background-color: var(--color-black);
+}
