+++ src/components/mainPage/VideoInfo.module.scss
@@ -20,10 +20,10 @@
   overflow: hidden;
   white-space: normal;
   line-height: 19px;
-  color: black;
+  color: var(--color-white);
 }
 .infochan_view {
   font-size: 13px;
-  color: gray;
+  color: var(--color-light-gray);
   margin: 8px 3px;
 }
