+++ src/components/mainPage/VideoMenu.module.scss
@@ -4,10 +4,9 @@
   width: 220px;
   right: 0;
   bottom: -10px;
-  border: 1px solid lightgray;
-  background-color: white;
-  box-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+  background-color: var(--color-dark-gray);
   z-index: 2;
+  border-radius: 10px;
 }
 .menulist {
   height: 36px;
@@ -16,7 +15,8 @@
 }
 .menubutton {
   display: flex;
+  color: var(--color-white);
 }
 .menulist:hover {
-  background-color: lightgray;
+  background-color: var(--color-semi-dark-gray);
 }
