+++ src/global.css
@@ -10,4 +10,5 @@ body {
   --color-light-gray: #aaaaaa;
   --color-semi-dark-gray: #323232;
   --color-dark-gray: #202020;
+  --color-semi-light-gray: #495057;
 }
